% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_well_tibble.R
\name{add_aer_status}
\alias{add_aer_status}
\title{A helper function for create_well_tibble(). Left joins the status data to the tibble.}
\usage{
add_aer_status(well_list, long_status_description)
}
\arguments{
\item{well_list}{A tibble containing the well_list with status codes}

\item{long_status_description}{A Boolean. Should the longer status
descriptions provided by the AER be used?}
}
\value{
The well tibble with additional columns containing the components of
  the well status
}
\description{
A helper function for create_well_tibble(). Left joins the status data to the tibble.
}
