% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_well_tibble.R
\name{add_business_associates}
\alias{add_business_associates}
\title{A helper function for create_well_tibble(). Left joins the business associate data to the tibble.}
\usage{
add_business_associates(well_list)
}
\arguments{
\item{well_list}{A tibble containing the well_list with business associate codes}
}
\value{
The well tibble with additional columns containing the licensee,
  agent, and operator data.
}
\description{
A helper function for create_well_tibble(). Left joins the business associate data to the tibble.
}
