% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_well_tibble.R
\name{add_field}
\alias{add_field}
\title{A helper function for create_well_tibble(). Left joins the field data to the tibble.}
\usage{
add_field(well_list)
}
\arguments{
\item{well_list}{A tibble containing the well_list with field codes}
}
\value{
The well tibble with additional columns containing the field and field centre
}
\description{
A helper function for create_well_tibble(). Left joins the field data to the tibble.
}
