% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_well_tibble.R
\name{add_oilsands_area_deposit}
\alias{add_oilsands_area_deposit}
\title{A helper function for create_well_tibble(). Left joins the oilsands area and deposit data to the tibble.}
\usage{
add_oilsands_area_deposit(well_list)
}
\arguments{
\item{well_list}{A tibble containing the well_list with os area and deposit codes}
}
\value{
The well tibble with additional columns containing the oilsands area and deposit
}
\description{
A helper function for create_well_tibble(). Left joins the oilsands area and deposit data to the tibble.
}
