% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_well_tibble.R
\name{add_pool}
\alias{add_pool}
\title{A helper function for create_well_tibble(). Left joins the pool data to the tibble.}
\usage{
add_pool(well_list)
}
\arguments{
\item{well_list}{A tibble containing the well_list with field codes}
}
\value{
The well tibble with additional columns containing the pool and confidential flag
}
\description{
A helper function for create_well_tibble(). Left joins the pool data to the tibble.
}
