% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_well_tibble.R
\name{convert_aer_dates}
\alias{convert_aer_dates}
\title{A helper function for create_well_tibble(). Converts the provided date strings to real dates}
\usage{
convert_aer_dates(well_list)
}
\arguments{
\item{well_list}{A tibble containing the well_list with unconverted dates}
}
\value{
The well tibble with additional columns containing the properly converted dates
}
\description{
A helper function for create_well_tibble(). Converts the provided date strings to real dates
}
