% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_well_tibble.R
\name{create_well_tibble}
\alias{create_well_tibble}
\title{Create a tibble conatining the official AER well list data.}
\usage{
create_well_tibble(remove_code_cols = TRUE, remove_original_dates = TRUE,
  long_status_description = FALSE, remove_ba_contacts = TRUE)
}
\arguments{
\item{remove_code_cols}{A Boolean. Should the original code columns be
removed from the final tibble?}

\item{remove_original_dates}{A Boolean. Should the the columns containing the
original dates as character vectors be removed?.}

\item{long_status_description}{A Boolean. Should the longer status
descriptions provided by the AER be used?}

\item{remove_ba_contacts}{A Boolean. Should the additional business associate
contact info such as address and phone number be removed?}
}
\value{
A tibble of the well list data with all the supporting data joined to
  the appropriate records.
}
\description{
Create a tibble conatining the official AER well list data.
}
\examples{
create_well_tibble()
create_well_tibble(remove_code_cols = FALSE, remove_original_dates = FALSE, long_status_description = TRUE, remove_ba_contacts = FALSE)
}
