% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_aer_data.R
\name{download_aer_data}
\alias{download_aer_data}
\title{Attempt to download the the missing files to the specified directory.}
\usage{
download_aer_data(data_dir = "data/")
}
\arguments{
\item{data_dir}{A path to a directory where the necessary AER data files should be saved.}
}
\description{
Attempt to download the the missing files to the specified directory.
}
\examples{
download_aer_data()
download_aer_data(data_dir = "some/other/directory/")
}
