% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_well_tibble.R
\name{remove_ba_contact_columns}
\alias{remove_ba_contact_columns}
\title{A helper function for create_well_tibble(). Remove the business associate contact columns.}
\usage{
remove_ba_contact_columns(well_list)
}
\arguments{
\item{well_list}{A tibble containing the expanded well_list}
}
\value{
The well tibble with business associate contact information removed
}
\description{
A helper function for create_well_tibble(). Remove the business associate contact columns.
}
