% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_well_tibble.R
\name{remove_code_columns}
\alias{remove_code_columns}
\title{A helper function for create_well_tibble(). Remove the code columns used to join status, field, pool etc.}
\usage{
remove_code_columns(well_list)
}
\arguments{
\item{well_list}{A tibble containing the expanded well_list}
}
\value{
The well tibble with all code columns removed.
}
\description{
A helper function for create_well_tibble(). Remove the code columns used to join status, field, pool etc.
}
