% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_well_tibble.R
\name{remove_original_date_columns}
\alias{remove_original_date_columns}
\title{A helper function for create_well_tibble(). Remove the unconverted original date columns.}
\usage{
remove_original_date_columns(well_list)
}
\arguments{
\item{well_list}{A tibble containing the expanded well_list}
}
\value{
The well tibble with original date columns removed
}
\description{
A helper function for create_well_tibble(). Remove the unconverted original date columns.
}
