% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_well_tibble.R
\name{reorder_columns}
\alias{reorder_columns}
\title{A helper function for create_well_tibble(). Reorder the columns in the aer well tibble to match the order provided in WellList.txt}
\usage{
reorder_columns(well_list)
}
\arguments{
\item{well_list}{A tibble containing the well_list with field codes}
}
\value{
The well tibble with columns that follow the orginal order
}
\description{
A helper function for create_well_tibble(). Reorder the columns in the aer well tibble to match the order provided in WellList.txt
}
